requirements.txt
mcp[cli]>=1.4.0
uvicorn>=0.29.0
starlette>=0.36.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
